+++ packages/material/src/PaginationItem/index.tsx
@@ -1,0 +1,7 @@
+export { default } from "./PaginationItem";
+export * from "./PaginationItem";
+
+export { default as paginationItemClasses } from "./paginationItemClasses";
+export * from "./paginationItemClasses";
+
+export * from "./PaginationItemProps";