+++ packages/site/src/pages/components/PaginationPage/PaginationButtonsExample.tsx
@@ -1,0 +1,10 @@
+import { Pagination, Stack } from "@suid/material";
+
+export default function PaginationButtons() {
+  return (
+    <Stack spacing={2}>
+      <Pagination count={10} showFirstButton showLastButton />
+      <Pagination count={10} hidePrevButton hideNextButton />
+    </Stack>
+  );
+}