+++ packages/site/src/pages/components/PaginationPage/PaginationOutlinedExample.tsx
@@ -1,0 +1,12 @@
+import { Pagination, Stack } from "@suid/material";
+
+export default function PaginationOutlined() {
+  return (
+    <Stack spacing={2}>
+      <Pagination count={10} variant="outlined" />
+      <Pagination count={10} variant="outlined" color="primary" />
+      <Pagination count={10} variant="outlined" color="secondary" />
+      <Pagination count={10} variant="outlined" disabled />
+    </Stack>
+  );
+}