+++ packages/site/src/pages/components/PaginationPage/PaginationRangesExample.tsx
@@ -1,0 +1,17 @@
+import { Pagination, Stack } from "@suid/material";
+
+export default function PaginationRanges() {
+  return (
+    <Stack spacing={2}>
+      <Pagination count={11} defaultPage={6} siblingCount={0} />
+      <Pagination count={11} defaultPage={6} /> {/* Default ranges */}
+      <Pagination
+        count={11}
+        defaultPage={6}
+        siblingCount={0}
+        boundaryCount={2}
+      />
+      <Pagination count={11} defaultPage={6} boundaryCount={2} />
+    </Stack>
+  );
+}