+++ packages/site/src/pages/components/PaginationPage/PaginationRoundedExample.tsx
@@ -1,0 +1,10 @@
+import { Pagination, Stack } from "@suid/material";
+
+export default function PaginationRounded() {
+  return (
+    <Stack spacing={2}>
+      <Pagination count={10} shape="rounded" />
+      <Pagination count={10} variant="outlined" shape="rounded" />
+    </Stack>
+  );
+}