+++ packages/site/src/pages/components/PaginationPage/PaginationSizeExample.tsx
@@ -1,0 +1,11 @@
+import { Pagination, Stack } from "@suid/material";
+
+export default function PaginationSize() {
+  return (
+    <Stack spacing={2}>
+      <Pagination count={10} size="small" />
+      <Pagination count={10} />
+      <Pagination count={10} size="large" />
+    </Stack>
+  );
+}