+++ packages/system/src/inspect.ts
@@ -1,4 +1,11 @@
-import { Component, createMemo, JSX } from "solid-js";
+import {
+  children,
+  Component,
+  createComponent,
+  createContext,
+  JSX,
+  useContext,
+} from "solid-js";
 
 export const $INSPECT = Symbol("solid-inspect");
 
@@ -13,21 +20,22 @@ type PrivateComponentObject<T = any> = ComponentObject<T> & {
 
 export type InspectResult = JSX.Element | ComponentObject;
 
-let inspectionEnabled = false;
+export const InspectContext = createContext<{ enabled: boolean }>();
 
-export function inspect(fn: () => JSX.Element): InspectResult[] {
-  try {
-    inspectionEnabled = true;
-    const result = fn();
-    return Array.isArray(result) ? result : [result];
-  } finally {
-    inspectionEnabled = false;
-  }
+export function inspectChildren(fn: () => JSX.Element): () => InspectResult[] {
+  const cb = createComponent(InspectContext.Provider, {
+    value: { enabled: true },
+    get children() {
+      return fn();
+    },
+  }) as any;
+  const result = children(cb);
+  return result.toArray;
 }
 
 export function componentTrap<T>(fn: Component<T>): Component<T> {
   function Component(props: T) {
-    if (inspectionEnabled)
+    if (useContext(InspectContext)?.enabled)
       return {
         Component,
         props,
@@ -53,33 +61,4 @@ export function isComponentObject(input: unknown, component?: Component) {
     (input as PrivateComponentObject).$INSPECT === $INSPECT &&
     (!component || (input as PrivateComponentObject).Component === component)
   );
-}
-
-export function resolveChildren(children: any): unknown {
-  if (typeof children === "function" && !children.length)
-    return resolveChildren(children());
-  if (Array.isArray(children)) {
-    const results: any[] = [];
-    for (let i = 0; i < children.length; i++) {
-      const result = resolveChildren(children[i]);
-      Array.isArray(result)
-        ? // eslint-disable-next-line prefer-spread
-          results.push.apply(results, result)
-        : results.push(result);
-    }
-    return results;
-  }
-  return children;
-}
-
-export function inspectChildren(fn: () => any) {
-  const children = createMemo(() => inspect(fn));
-  const memo = createMemo(() =>
-    inspect(() => resolveChildren(children()) as any)
-  );
-  (memo as any).toArray = () => {
-    const c = memo();
-    return Array.isArray(c) ? c : c != null ? [c] : [];
-  };
-  return memo;
 }