+++ packages/icons-material/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @suid/icons-material
 
+## 0.8.1
+
+### Patch Changes
+
+- [`b58930c`](https://github.com/swordev/suid/commit/b58930ca80080062f5a9f2a9bcf42fff9a623e53) Thanks [@juanrgm](https://github.com/juanrgm)! - Update dependencies
+
+- Updated dependencies [[`6406d78`](https://github.com/swordev/suid/commit/6406d78845ef694b044c36a7739edf823c6f286a), [`b58930c`](https://github.com/swordev/suid/commit/b58930ca80080062f5a9f2a9bcf42fff9a623e53), [`7f086b2`](https://github.com/swordev/suid/commit/7f086b2d3de454426147ea182a69b26c24ce09d2)]:
+  - @suid/material@0.18.0
+
+## 0.8.0
+
+### Minor Changes
+
+- [`bee2e07`](https://github.com/swordev/suid/commit/bee2e07ac95a9042e2971cb54c5ddcf45018c959) Thanks [@juanrgm](https://github.com/juanrgm)! - Update dependencies
+
+### Patch Changes
+
+- Updated dependencies [[`bee2e07`](https://github.com/swordev/suid/commit/bee2e07ac95a9042e2971cb54c5ddcf45018c959), [`8ea405f`](https://github.com/swordev/suid/commit/8ea405f9754ac02eee855d6c1e06721f4df012c8)]:
+  - @suid/material@0.17.0
+
 ## 0.7.0
 
 ### Minor Changes