+++ packages/material/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @suid/material
 
+## 0.18.0
+
+### Minor Changes
+
+- [`6406d78`](https://github.com/swordev/suid/commit/6406d78845ef694b044c36a7739edf823c6f286a) Thanks [@juanrgm](https://github.com/juanrgm)! - Resolve the theme context on access
+
+- [`7f086b2`](https://github.com/swordev/suid/commit/7f086b2d3de454426147ea182a69b26c24ce09d2) Thanks [@juanrgm](https://github.com/juanrgm)! - Add the `Pagination` and `PaginationItem` components
+
+### Patch Changes
+
+- [`b58930c`](https://github.com/swordev/suid/commit/b58930ca80080062f5a9f2a9bcf42fff9a623e53) Thanks [@juanrgm](https://github.com/juanrgm)! - Update dependencies
+
+- Updated dependencies [[`6406d78`](https://github.com/swordev/suid/commit/6406d78845ef694b044c36a7739edf823c6f286a), [`b196443`](https://github.com/swordev/suid/commit/b196443f4f0ba0b962ad0438bb1db059b5755d19), [`b58930c`](https://github.com/swordev/suid/commit/b58930ca80080062f5a9f2a9bcf42fff9a623e53)]:
+  - @suid/system@0.13.0
+  - @suid/types@0.7.1
+  - @suid/utils@0.10.1
+  - @suid/base@0.10.1
+
+## 0.17.0
+
+### Minor Changes
+
+- [`bee2e07`](https://github.com/swordev/suid/commit/bee2e07ac95a9042e2971cb54c5ddcf45018c959) Thanks [@juanrgm](https://github.com/juanrgm)! - Update dependencies
+
+### Patch Changes
+
+- [`8ea405f`](https://github.com/swordev/suid/commit/8ea405f9754ac02eee855d6c1e06721f4df012c8) Thanks [@juanrgm](https://github.com/juanrgm)! - Fix circular dependency
+
+- Updated dependencies [[`bee2e07`](https://github.com/swordev/suid/commit/bee2e07ac95a9042e2971cb54c5ddcf45018c959), [`55b070c`](https://github.com/swordev/suid/commit/55b070ceb72a90d8fb9de85d8c8deca287a2ca80)]:
+  - @suid/system@0.12.0
+  - @suid/types@0.7.0
+  - @suid/utils@0.10.0
+  - @suid/base@0.10.0
+
 ## 0.16.0
 
 ### Minor Changes